# openapi.yaml
openapi: 3.1.0
info:
  title: Songless Multiplayer API
  version: 0.1.0
  description: |
    REST + WebSocket API for the multiplayer Songless variant.

    ## Auth model
    All non-auth endpoints require a session cookie (`session=<token>`).
    The cookie is set by `POST /auth/login` and cleared by `POST /auth/logout`.

    ## Gameplay model (Heardle-style)
    Each round, every non-picker player progresses through guess brackets
    independently. At bracket index `i`, the player can play up to
    `guess_brackets_seconds[i]` seconds of audio. Each attempt the player
    either:
      - Guesses correctly → round ends for that player; they score
        `len(brackets) - i` points (default scale: 6/5/4/3/2/1).
      - Guesses incorrectly → bracket auto-advances by one.
      - Skips → bracket auto-advances by one (no guess recorded).
    A player is "exhausted" when they wrong-guess or skip past the last
    bracket — they finish the round with 0 points. The round ends when every
    non-picker has either guessed correctly or exhausted brackets, or when
    a hard server-side timeout elapses.

    ## Realtime (WebSocket: `/rooms/{code}/ws`)
    Server → client events (`{ "type": ..., "payload": ... }`):
      - `room_state`           — full snapshot, sent on connect / after any change
      - `player_joined`        — `{ user }`
      - `player_left`          — `{ user_id }`
      - `player_disconnected`  — `{ user_id }`
      - `settings_updated`     — `{ settings }`
...

...

...
servers:
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
    User:
      type: object
      required: [id, username]
      properties:
        id: { type: string, format: uuid }
        username: { type: string }
    RoomStatus:
      type: string
      enum: [lobby, selecting, playing, results]
    RoomSettings:
      type: object
      properties:
        min_popularity:
          type: integer
          minimum: 0
          maximum: 100
          description: Minimum Spotify popularity (0-100).
        required_artists:
          type: array
          items: { type: string }
          description: Spotify artist IDs. If non-empty, track must be by one of these.
        songs_per_player:
          type: integer
          minimum: 1
          maximum: 10
        guess_brackets_seconds:
          type: array
          items: { type: number, minimum: 0 }
          minItems: 1
          description: |
            Strictly-ascending elapsed-time markers (in seconds) at which the score
            decreases. The last value is also the round's hard time limit.
            Default: [0.5, 1, 2.5, 5, 15, 30].
        album_art_enabled:
          type: boolean
          description: Show the song's album cover during the round.
        album_art_unblur:
          type: boolean
          description: |
            When enabled, the album cover is blurred at round start and progressively
            unblurred as each guess bracket elapses. Has no effect if
            `album_art_enabled` is false.
        round_intermission_seconds:
          type: integer
          minimum: 0
          maximum: 30
          description: |
            Pause between rounds during which the song is revealed and the
            leaderboard movement is animated. 0 disables the intermission.
        post_game_delay_seconds:
          type: integer
          minimum: 0
          description: Cooldown after results before leader can restart.
    RoomPlayer:
      type: object
      required: [user, score, connected, songs_submitted]
      properties:
        user: { $ref: "#/components/schemas/User" }
        score: { type: integer }
        connected: { type: boolean }
        songs_submitted:
          {
            type: integer,
            description: Count only; titles withheld until reveal.,
          }
    Room:
      type: object
      required: [code, leader_id, status, settings, players]
      properties:
        code: { type: string }
        leader_id: { type: string, format: uuid }
        status: { $ref: "#/components/schemas/RoomStatus" }
        settings: { $ref: "#/components/schemas/RoomSettings" }
        players:
          type: array
          items: { $ref: "#/components/schemas/RoomPlayer" }
        current_round_id:
          type: string
          format: uuid
          nullable: true
...
    SubmittedSong:
      type: object
      required: [id, spotify_track_id, title, artist]
      properties:
        id: { type: string, format: uuid }
        spotify_track_id: { type: string }
        title: { type: string }
        artist: { type: string }
        preview_url: { type: string, nullable: true }
        album_image_url: { type: string, nullable: true }
...
    ScoreboardEntry:
      type: object
      required: [user, score]
      properties:
        user: { $ref: "#/components/schemas/User" }
        score: { type: integer }
        previous_score:
          type: integer
          description: |
            Score before this round's points were applied. Only present in the
            `round_ended` WS event so clients can animate rank changes.
  responses:
    Unauthorized:
      description: Not logged in
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Authenticated but not allowed (e.g. not leader)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Room or resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: Action conflicts with current room state
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
security:
  - sessionCookie: []
paths:
...
  /rooms:
    post:
      summary: Create a new room (caller becomes leader)
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Room" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /rooms/{code}:
    parameters:
      - { name: code, in: path, required: true, schema: { type: string } }
    get:
      summary: Get room state
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Room" }
        "404": { $ref: "#/components/responses/NotFound" }
  /rooms/{code}/join:
    parameters:
      - { name: code, in: path, required: true, schema: { type: string } }
    post:
      summary: Join an existing room
      responses:
        "200":
          description: Joined
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Room" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Room is no longer in lobby phase
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /rooms/{code}/leave:
    parameters:
      - { name: code, in: path, required: true, schema: { type: string } }
    post:
      summary: Leave a room
      responses:
        "204": { description: Left }
        "404": { $ref: "#/components/responses/NotFound" }
  /rooms/{code}/settings:
    parameters:
      - { name: code, in: path, required: true, schema: { type: string } }
    patch:
      summary: Update room settings (leader only, lobby phase only)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RoomSettings" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Room" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
  /rooms/{code}/phase:
    parameters:
      - { name: code, in: path, required: true, schema: { type: string } }
    post:
      summary: Advance the room to the next phase (leader only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [target]
              properties:
                target: { $ref: "#/components/schemas/RoomStatus" }
      responses:
        "200":
          description: Phase changed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Room" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409":
          description: Illegal transition (e.g. starting game before all songs submitted)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /rooms/{code}/restart:
    parameters:
      - { name: code, in: path, required: true, schema: { type: string } }
    post:
      summary: Reset scores and songs, return to lobby (leader only, results phase only)
      responses:
        "200":
          description: Restarted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Room" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
  /rooms/{code}/songs:
    parameters:
      - { name: code, in: path, required: true, schema: { type: string } }
    get:
      summary: My submitted songs for this room
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SubmittedSong" }
    post:
      summary: Submit a song. Server re-validates against room rules via Spotify.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [spotify_track_id]
              properties:
                spotify_track_id: { type: string }
      responses:
        "201":
          description: Accepted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SubmittedSong" }
        "409":
          description: Already submitted, or quota reached
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422":
          description: Track does not match room rules (popularity / genre / artist / no preview_url)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /rooms/{code}/songs/{song_id}:
    parameters:
      - { name: code, in: path, required: true, schema: { type: string } }
      - {
          name: song_id,
          in: path,
          required: true,
          schema: { type: string, format: uuid },
        }
    delete:
      summary: Remove one of my own songs (selection phase only)
      responses:
        "204": { description: Removed }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
...
  /rooms/{code}/results:
    parameters:
      - { name: code, in: path, required: true, schema: { type: string } }
    get:
      summary: Final scoreboard (results phase)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ScoreboardEntry" }
        "409": { $ref: "#/components/responses/Conflict" }
...